# frontend/apps/admin/openapi/7dpanel.v1.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "7DPanel API",
    "description": "Runtime API for 7DPanel 0.1.0.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/api/v1/access-lists/bans": {
      "get": {
        "tags": [
          "AccessLists"
        ],
        "operationId": "AccessLists_GetBans",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BanEntryHttpResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/access-lists/whitelist": {
      "get": {
        "tags": [
          "AccessLists"
        ],
        "operationId": "AccessLists_GetWhitelist",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WhitelistEntryHttpResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/access-lists/bans/{playerId}": {
      "put": {
        "tags": [
          "AccessLists"
        ],
        "operationId": "AccessLists_PutBan",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/BanUpsertHttpRequest"
                  }
                ]
              }
            }
          },
          "x-position": 2
        },
        "responses": {
          "204": {
            "description": "The operation completed successfully."
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "AccessLists"
        ],
        "operationId": "AccessLists_DeleteBan",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": "The operation completed successfully."
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/access-lists/whitelist/{playerId}": {
      "put": {
        "tags": [
          "AccessLists"
        ],
        "operationId": "AccessLists_PutWhitelist",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/WhitelistUpsertHttpRequest"
                  }
                ]
              }
            }
          },
          "x-position": 2
        },
        "responses": {
          "204": {
            "description": "The operation completed successfully."
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "AccessLists"
        ],
        "operationId": "AccessLists_DeleteWhitelist",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": "The operation completed successfully."
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/v1/game-permissions/admins": {
      "get": {
        "tags": [
          "GamePermissions"
        ],
        "operationId": "GamePermissions_GetAdmins",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameAdminHttpResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/game-permissions/commands": {
      "get": {
        "tags": [
          "GamePermissions"
        ],
        "operationId": "GamePermissions_GetCommands",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommandPermissionHttpResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/game-permissions/admins/{playerId}": {
      "put": {
        "tags": [
          "GamePermissions"
        ],
        "operationId": "GamePermissions_PutAdmin",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/GameAdminUpsertHttpRequest"
                  }
                ]
              }
            }
          },
          "x-position": 2
        },
        "responses": {
          "204": {
            "description": "The operation completed successfully."
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "GamePermissions"
        ],
        "operationId": "GamePermissions_DeleteAdmin",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": "The operation completed successfully."
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/game-permissions/commands/{command}": {
      "put": {
        "tags": [
          "GamePermissions"
        ],
        "operationId": "GamePermissions_PutCommand",
        "parameters": [
          {
            "name": "command",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CommandPermissionUpsertHttpRequest"
                  }
                ]
              }
            }
          },
          "x-position": 2
        },
        "responses": {
          "204": {
            "description": "The operation completed successfully."
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "GamePermissions"
        ],
        "operationId": "GamePermissions_DeleteCommand",
        "parameters": [
          {
            "name": "command",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": "The operation completed successfully."
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "BanEntryHttpResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "playerId": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "bannedUntilUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "WhitelistEntryHttpResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "playerId": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        }
      },
      "BanUpsertHttpRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "bannedUntilUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "WhitelistUpsertHttpRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "GameAdminHttpResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "playerId": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "permissionLevel": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CommandPermissionHttpResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "command": {
            "type": "string"
          },
          "permissionLevel": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "GameAdminUpsertHttpRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "permissionLevel": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CommandPermissionUpsertHttpRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "permissionLevel": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Opaque Bearer access token.",
        "scheme": "bearer",
        "bearerFormat": "opaque"
      }
    }
  }
}